Remove useUsers dependency and offload to useTestUser temporarily
